refactor: UserProfileResponse 분리
