--- DevLog/Data/DTO/UserProfile+.swift
@@ -1,24 +1,0 @@
-//
-//  UserProfile+.swift
-//  DevLog
-//
-//  Created by 최윤진 on 1/10/26.
-//
-
-import Foundation
-
-struct UserProfileResponse: Decodable {
-    let name: String
-    let email: String
-    let statusMessage: String
-    let avatarURL: URL?
-
-    func toDomain() -> UserProfile {
-        return UserProfile(
-            name: self.name,
-            email: self.email,
-            statusMessage: self.statusMessage,
-            avatarURL: self.avatarURL
-        )
-    }
-}
